Fix crash due to TomlConfigHandler.getFloat not working because it reads float config value as a double.

+++ src/main/java/teamport/aether/AetherConfig.java
@@ -120,7 +120,7 @@ public class AetherConfig {
                 res = (T) new Long(cfg.getLong(key));
             }
             else if (def instanceof Float) {
-                res = (T) new Float(cfg.getFloat(key));
+                res = (T) new Float(cfg.getDouble(key));
             }
             else if (def instanceof Double) {
                 res = (T) new Double(cfg.getDouble(key));
